+++ src/main/java/life/majiang/community/model/Question.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class Question {
-    private Integer id;
+    private Long id;
 
     private String title;
 
@@ -12,7 +12,7 @@ public class Question {
 
     private Long gmtModified;
 
-    private Integer creator;
+    private Long creator;
 
     private Integer commentCount;
 
@@ -23,5 +23,4 @@ public class Question {
     private String tag;
 
     private String description;
-
 }
